Separate out default custom element binding.
