+++ src/api/type.js
@@ -1,3 +1,5 @@
+import binding from '../polyfill/binding';
+
 function getClassList (element) {
   var classList = element.classList;
 
@@ -65,34 +67,5 @@ export default {
       return definitions;
     }
   },
-  ELEMENT: {
-    /* jshint expr: true */
-    create (opts) {
-      var elem = document.createElement(opts.extends || opts.id);
-      opts.extends && elem.setAttribute('is', opts.id);
-      return elem;
-    },
-    find (elem, defs) {
-      var attrs = elem.attributes;
-      var definitions = [];
-      var isAttr = attrs.is;
-      var isAttrValue = isAttr && (isAttr.value || isAttr.nodeValue);
-      var tag = elem.tagName.toLowerCase();
-      var definition = defs[isAttrValue || tag];
-
-      if (definition && definition.type === this) {
-        let tagToExtend = definition.extends;
-
-        if (isAttrValue) {
-          if (tag === tagToExtend) {
-            definitions.push(definition);
-          }
-        } else if (!tagToExtend) {
-          definitions.push(definition);
-        }
-      }
-
-      return definitions;
-    }
-  }
+  ELEMENT: binding
 };
